+++ example/lib/src/screens/features/app_bar_example.dart
@@ -50,6 +50,7 @@ class AppBarExample extends StatelessWidget {
                   ),
                   showSearch: false,
                   showMenu: false,
+                  onTitleTap: () => _showSnackBar(context, 'Title tapped'),
                 ),
               ),
             ),